component editor - extract common components
